LayoutTests/imported/w3c/web-platform-tests/cors/status-expected.txt

CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?1&origin=none&code=400 due to access control checks.
CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?5&headers=x-custom&preflight=400 due to access control checks.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/cors-makeheader.py?2&origin=none&preflight=200 due to access control checks.
The returned status code in different scenarios

   allowed  preflight  response  | status |
   -------  ---------  --------  | ------ |
 1      no          x       400  |      0 |
 2      no        200         x  |      0 |
 3     yes          x       400  |    400 |
 4     yes        200       400  |    400 |
 5     yes        400         x  |      0 |

PASS 1. CORS disallowed, response status 400. 
PASS 2. CORS disallowed, preflight status 200. 
PASS 3. CORS allowed, response status 400. 
PASS 4. CORS allowed, preflight status 200, response status 400. 
PASS 5. CORS allowed, preflight status 400. 
